Create interface class for Game.
Update all references to the old class and fix errors related
to how the new class is defined.
